fix: added filter dropdown on mobile position page
